+++ security-analysis/security-analysis-api/src/main/java/com/powsybl/security/limitreduction/ReducedApparentPowerLimits.java
@@ -16,7 +16,7 @@ import com.powsybl.iidm.network.ApparentPowerLimits;
  */
 public class ReducedApparentPowerLimits extends AbstractReducedLoadingLimits implements ApparentPowerLimits {
     public ReducedApparentPowerLimits(double permanentLimit, double originalPermanentLimit,
-                                      double reductionAppliedOnPermanentLimit) {
+                                      float reductionAppliedOnPermanentLimit) {
         super(permanentLimit, originalPermanentLimit, reductionAppliedOnPermanentLimit);
     }
 }
